+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 ------------------
 
 - fix crash on export when model has no ``id`` (`1802 <https://github.com/django-import-export/django-import-export/pull/1802>`_)
+- fix Widget crash when django Field subclass is used (`1805 <https://github.com/django-import-export/django-import-export/pull/1805>`_)
 - fix regression: allow imports when default ``import_id_field`` is not present (`1813 <https://github.com/django-import-export/django-import-export/pull/1813>`_)
 
 4.0.0 (2024-04-27)